refactor: migrate slash command handlers to topic-specific files (T012)
Distribute ~30 slash command handler methods from rpc_app.go into 4
topic-specific files:

- rpc_session_handlers.go: /new, /resume, /session, /rewind, /fork,
  hidden session commands
- rpc_message_handlers.go: /messages, /compact, /export_html,
  hidden message commands, getWorkflowStatus, formatMessagesForDisplay
- rpc_config_handlers.go: /set, /model, /thinking, /toggle, /show,
  /context, /trace-events, hidden config commands
- rpc_help_handlers.go: /help, /skills, /quit, /abort, /follow-up,
  hidden aliases

The registerHandlers method now delegates to topic-specific
registerXxxHandlers() methods. rpc_app.go reduced from 2008 to 717
lines (struct, init, infrastructure, protocol handlers only).
rpc_handlers.go reduced from 599 to 443 lines (runRPC only).
