+++ cmd/ai/rpc_handlers.go
@@ -1,15 +1,11 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
-	agentctx "github.com/tiancaiamao/ai/pkg/context"
 	"io"
 	_ "net/http/pprof"
 	"os"
 	"path/filepath"
-	"regexp"
-	"strings"
 	"time"
 
 	"log/slog"
@@ -445,155 +441,3 @@ func runRPC(sessionPath string, debugAddr string, input io.Reader, output io.Wri
 }
 
 // getWorkflowStatus reads .workflow/state.json and tasks.md to return workflow state.
-func getWorkflowStatus(cwd string) (*rpc.WorkflowState, error) {
-	state := &rpc.WorkflowState{
-		Phase:      "not_started",
-		LastUpdate: time.Now().UTC().Format(time.RFC3339),
-	}
-
-	workflowDir := filepath.Join(cwd, ".workflow")
-	stateFile := filepath.Join(workflowDir, "state.json")
-
-	// Read state.json if it exists
-	if data, err := os.ReadFile(stateFile); err == nil {
-		var stateData struct {
-			Phase     string `json:"phase"`
-			StartedAt string `json:"started_at"`
-			TasksFile string `json:"tasks_file"`
-		}
-		if err := json.Unmarshal(data, &stateData); err == nil {
-			state.Phase = stateData.Phase
-			state.StartedAt = stateData.StartedAt
-			if stateData.TasksFile != "" {
-				// Handle relative or absolute path
-				if filepath.IsAbs(stateData.TasksFile) {
-					state.TasksFile = stateData.TasksFile
-				} else {
-					state.TasksFile = filepath.Join(cwd, stateData.TasksFile)
-				}
-			}
-		}
-	}
-
-	// Read tasks.md if specified
-	if state.TasksFile != "" {
-		tasksData, err := os.ReadFile(state.TasksFile)
-		if err != nil {
-			return nil, fmt.Errorf("failed to read tasks file %s: %w", state.TasksFile, err)
-		}
-
-		// Parse task statuses
-		lines := strings.Split(string(tasksData), "\n")
-		var inProgressTask *rpc.WorkflowTask
-
-		for _, line := range lines {
-			line = strings.TrimSpace(line)
-			if !strings.HasPrefix(line, "- [") {
-				continue
-			}
-
-			status := "pending"
-			if strings.HasPrefix(line, "- [x]") || strings.HasPrefix(line, "- [X]") {
-				status = "done"
-				state.DoneTasks++
-			} else if strings.HasPrefix(line, "- [-]") {
-				status = "in_progress"
-				state.PendingTasks++ // In-progress also counts toward pending
-			} else if strings.HasPrefix(line, "- [!]") {
-				status = "failed"
-				state.FailedTasks++
-			} else {
-				state.PendingTasks++
-			}
-
-			state.TotalTasks++
-
-			// Extract task ID and description for in-progress task
-			if status == "in_progress" && inProgressTask == nil {
-				// Extract task ID (e.g., TASK001, T01, etc.)
-				var id string
-				idMatch := regexp.MustCompile(`[A-Z]{3,}\d+|[A-Z]\d+`).FindString(line)
-				if idMatch != "" {
-					id = idMatch
-				}
-
-				// Extract description: remove checkbox first, then task ID
-				desc := line
-				// Remove checkbox
-				desc = regexp.MustCompile(`^-\s*\[[xX\-\!]\]\s*`).ReplaceAllString(desc, "")
-				desc = regexp.MustCompile(`^-\s*\[\s*\]\s*`).ReplaceAllString(desc, "")
-				// Remove task ID (e.g., TASK002: or TASK002 )
-				desc = regexp.MustCompile(`^[A-Z]{3,}\d+:?\s*`).ReplaceAllString(desc, "")
-				desc = strings.TrimSpace(desc)
-
-				inProgressTask = &rpc.WorkflowTask{
-					ID:          id,
-					Description: desc,
-					Status:      status,
-				}
-			}
-		}
-
-		state.InProgressTask = inProgressTask
-	}
-
-	return state, nil
-}
-
-// formatMessagesForDisplay converts AgentMessages into a structured summary for the /messages command.
-// It returns the last `count` messages with previews truncated to maxPreviewLen characters.
-func formatMessagesForDisplay(messages []agentctx.AgentMessage, count int, maxPreviewLen int) rpc.MessagesResult {
-	total := len(messages)
-
-	start := total - count
-	if start < 0 {
-		start = 0
-	}
-	showing := total - start
-
-	formatted := make([]rpc.FormattedMessage, 0, showing)
-	for i := start; i < total; i++ {
-		msg := messages[i]
-		fm := rpc.FormattedMessage{
-			Index: i,
-			Role:  msg.Role,
-		}
-
-		// Build preview from text content
-		preview := msg.ExtractText()
-		if preview == "" {
-			// Try thinking content as fallback for assistant messages
-			if thinking := msg.ExtractThinking(); thinking != "" {
-				preview = "(thinking) " + thinking
-			}
-		}
-		if len(preview) > maxPreviewLen {
-			preview = preview[:maxPreviewLen] + "..."
-		}
-		fm.Preview = preview
-
-		// Extract tool call names for assistant messages
-		toolCalls := msg.ExtractToolCalls()
-		if len(toolCalls) > 0 {
-			names := make([]string, 0, len(toolCalls))
-			for _, tc := range toolCalls {
-				names = append(names, tc.Name)
-			}
-			fm.ToolCalls = names
-		}
-
-		// Include tool name for tool results
-		if msg.ToolName != "" {
-			fm.ToolName = msg.ToolName
-		}
-		fm.IsError = msg.IsError
-
-		formatted = append(formatted, fm)
-	}
-
-	return rpc.MessagesResult{
-		Total:    total,
-		Showing:  showing,
-		Messages: formatted,
-	}
-}
